+++ backend/seeders/04-orderItemSeeders.js
@@ -7,14 +7,34 @@ module.exports = {
         orderId: 1,
         productId: 1,
         quantity: 1,
-        petImageUrl: "...",
-        aiDesignImageUrl: "...",
+        petImageUrl: "http://localhost:3000/images/clients/dog1-original.jpg",
+        aiDesignImageUrl: "http://localhost:3000/images/catalog/products/pillow-design-dog3.png",
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      },
+      {
+        orderId: 2,
+        productId: 2,
+        quantity: 1,
+        petImageUrl: "http://localhost:3000/images/clients/dog2-original.jpg",
+        aiDesignImageUrl: "http://localhost:3000/images/catalog/products/bag-design-dog2.png",
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      },
+      {
+        orderId: 3,
+        productId: 3,
+        quantity: 2,
+        petImageUrl: "http://localhost:3000/images/clients/dog3-original.jpg",
+        aiDesignImageUrl: "http://localhost:3000/images/catalog/products/pillow-design-dog3.png",
         createdAt: new Date(),
         updatedAt: new Date(),
       },
     ]);
   },
   async down(queryInterface, Sequelize) {
-    await queryInterface.bulkDelete("OrderItems", null, {});
+    if (await queryInterface.tableExists("OrderItems")) {
+      await queryInterface.bulkDelete("OrderItems", null, {});
+    }
   },
 };
